+++ HISTORY.md
@@ -1,3 +1,4 @@
+*  Support for `many-to-many` relationships (Dennis Roberts)
 *  Fixed table-name delimiting, and at the same time, support Postgres' schema and 
    queries covering multiple databases.  
    See: https://github.com/korma/Korma/pull/105 (Tsutomu YANO)